+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [5.2.4](https://github.com/spipu/html2pdf/compare/v5.2.3...master) - unreleased
 
   * revert fix multibyte aware substr when setting newline position - it causes pbs on some specific cases
+  * add security on scheme of css and image paths 
 
 ## [5.2.3](https://github.com/spipu/html2pdf/compare/v5.2.2...v5.2.3) - 2021-10-19